integrating exercise3 Krystian

+++ PROD/Exercise3/Krystian.cs
@@ -19,7 +19,7 @@ namespace Exercise3
                     Prime++;
                 }
             }
-            return Prime;
+            return (int)Prime;
         }
     }
 }
